divide page ui with Ctrl-C&V
