+++ vj4/ui/pages/contest_main.page.styl
@@ -1,15 +1,6 @@
 $highlight-button-color = #F6DF45
 
 .page--contest_main
-  .homework__title
-    line-height: 1.2em
-    margin-bottom: rem(10px)
-    font-size: rem($font-size-title)
-
-  .homework__date
-    margin-right: rem(10px)
-    white-space: nowrap
-
   .contest__title
     font-size: rem($font-size-title)
     line-height: 1.2em
