+++ vj4/ui/pages/homework_main.page.styl
@@ -1,0 +1,11 @@
+$highlight-button-color = #F6DF45
+
+.page--homework_main
+  .homework__title
+    line-height: 1.2em
+    margin-bottom: rem(10px)
+    font-size: rem($font-size-title)
+
+  .homework__date
+    margin-right: rem(10px)
+    white-space: nowrap
