+++ vj4/ui/pages/homework_scoreboard.page.styl
@@ -1,0 +1,22 @@
+.page--homework_scoreboard
+  .col--rank
+    width: rem(70px)
+    border-right: 1px solid $table-border-color
+
+  .col--user
+    width: rem(140px)
+
+  .col--total_score, .col--total_time_str, .col--solved_problems
+    width: rem(100px)
+    text-align: center
+
+  .col--problem_detail
+    border-left: 1px solid $table-border-color
+    text-align: center
+
+  +mobile()
+    .col--user
+      width: auto
+
+    .col--problem_detail
+      display: none
